ref: trainer 1/n (#3412)
* ref: moved eval loop 2/n

* ref: moved eval loop 2/n

* ref: trainer 1/n

* ref: trainer 1/n

* ref: trainer 1/n
