+++ pytorch_lightning/trainer/initializer.py
@@ -12,22 +12,25 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from abc import ABC
-
 from pytorch_lightning import _logger as log
 from pytorch_lightning.utilities import APEX_AVAILABLE, NATIVE_AMP_AVALAIBLE, rank_zero_warn, AMPType
 
 
-class TrainerAMPMixin(ABC):
+class Initializer:
 
-    # this is just a summary on variables used in this abstract class,
-    #  the proper values/initialisation should be done in child class
-    precision: int
+    def __init__(self, trainer):
+        self.trainer = trainer
+
+    def init_amp(self, amp_type: str):
+        assert self.trainer.precision in (16, 32), 'only 32 or 16 bit precision supported'
+        self.trainer.amp_backend = None
+        self._setup_amp_backend(amp_type)
 
     def _setup_amp_backend(self, amp_type: str):
-        if self.precision != 16:
+        if self.trainer.precision != 16:
             # no AMP requested, so we can leave now
             return
+
         amp_type = amp_type.lower()
         assert amp_type in ('native', 'apex'), f'Unsupported amp type {amp_type}'
         if amp_type == 'native':
@@ -38,20 +41,16 @@ class TrainerAMPMixin(ABC):
                 amp_type = 'apex'
             else:
                 log.info('Using native 16bit precision.')
-                self.amp_backend = AMPType.NATIVE
+                self.trainer.amp_backend = AMPType.NATIVE
         if amp_type == 'apex':
             if not APEX_AVAILABLE:
                 rank_zero_warn('You have asked for Apex AMP but you have not installed it yet.'
                                ' Install apex first using this guide: https://github.com/NVIDIA/apex#linux')
             else:
                 log.info('Using APEX 16bit precision.')
-                self.amp_backend = AMPType.APEX
-        if not self.amp_backend:
+                self.trainer.amp_backend = AMPType.APEX
+        if not self.trainer.amp_backend:
             raise ModuleNotFoundError(
                 f'You have asked for AMP support {amp_type}, but there is no support on your side yet.'
                 f' Consider installing torch >= 1.6 or NVIDIA Apex.'
             )
-
-    @property
-    def use_amp(self) -> bool:
-        return self.precision == 16
